modifié :         BDD/BDD Demand.sql 	modifié :         catalogDemandMS/src/main/java/fr/insa/helperapp/catalogDemandMS/Services/demandService.java 	modifié :         catalogDemandMS/src/main/java/fr/insa/helperapp/catalogDemandMS/controler/demandManagement.java 	modifié :         catalogDemandMS/src/main/java/fr/insa/helperapp/catalogDemandMS/model/demandModel.java 	modifié :         statusDemandMS/src/main/java/fr/insa/helperapp/statusDemandMS/controller/StatusResource.java 	nouveau fichier : statusDemandMS/src/main/java/fr/insa/helperapp/statusDemandMS/model/AggregatedResponse.java 	modifié :         statusDemandMS/src/main/java/fr/insa/helperapp/statusDemandMS/model/Status.java 	modifié :         statusDemandMS/src/main/java/fr/insa/helperapp/statusDemandMS/repository/StatusRepository.java 	modifié :         statusDemandMS/src/main/resources/application.properties 	modifié :         userManagementMS/src/main/java/fr/insa/helperapp/userManagementMS/controller/UserResource.java 	nouveau fichier : whoIsTheResolverMS.properties 	nouveau fichier : whoIsTheResolverMS/.gitattributes 	nouveau fichier : whoIsTheResolverMS/.gitignore 	nouveau fichier : whoIsTheResolverMS/.mvn/wrapper/maven-wrapper.properties 	nouveau fichier : whoIsTheResolverMS/mvnw 	nouveau fichier : whoIsTheResolverMS/mvnw.cmd 	nouveau fichier : whoIsTheResolverMS/pom.xml 	nouveau fichier : whoIsTheResolverMS/src/main/java/fr/insa/helperapp/whoIsTheResolverMS/WhoIsTheResolverMsApplication.java 	nouveau fichier : whoIsTheResolverMS/src/main/java/fr/insa/helperapp/whoIsTheResolverMS/controller/ResolverResource.java 	nouveau fichier : whoIsTheResolverMS/src/main/java/fr/insa/helperapp/whoIsTheResolverMS/model/Aggregate.java 	nouveau fichier : whoIsTheResolverMS/src/main/java/fr/insa/helperapp/whoIsTheResolverMS/model/Resolver.java 	nouveau fichier : whoIsTheResolverMS/src/main/java/fr/insa/helperapp/whoIsTheResolverMS/repository/ResolverRepository.java 	nouveau fichier : whoIsTheResolverMS/src/main/resources/application.properties 	nouveau fichier : whoIsTheResolverMS/src/test/java/fr/insa/helperapp/whoIsTheResolverMS/WhoIsTheResolverMsApplicationTests.java

+++ catalogDemandMS/src/main/java/fr/insa/helperapp/catalogDemandMS/Services/demandService.java
@@ -24,7 +24,7 @@ public class demandService {
         @Override
         public demandModel mapRow(ResultSet rs, int rowNum) throws SQLException {
             return new demandModel(
-                    rs.getLong("id"),
+                    rs.getInt("id"),
                     rs.getString("name"),
                     rs.getString("description"),
                     rs.getString("author") 
@@ -37,7 +37,7 @@ public class demandService {
         return jdbcTemplate.query(sql, new DemandRowMapper());
     }
 
-    public demandModel getRequestById(Long id) {
+    public demandModel getRequestById(int id) {
         String sql = "SELECT * FROM demand WHERE id = ?";
         return jdbcTemplate.queryForObject(sql, new DemandRowMapper(), id);
     }
@@ -52,7 +52,7 @@ public class demandService {
         jdbcTemplate.update(sql, demand.getName(), demand.getDescription(), demand.getAuthor(), demand.getId());
     }
 
-    public void deleteRequest(Long id) {
+    public void deleteRequest(int id) {
         String sql = "DELETE FROM demand WHERE id = ?";
         jdbcTemplate.update(sql, id);
     }
